+++ src/resource/addProjectToSolution/AddProjectToSolution.ts
@@ -1,18 +1,17 @@
 import * as vscode from 'vscode';
-import * as parentFinder from 'find-parent-dir';
 import * as path from 'path';
 import * as fs from 'fs';
-import { CreateProjectPanel, ProjectState } from './CreateProject';
-const findUpGlob = require('find-up-glob');
+import { Panel } from './Panel';
+
 
 export class AddProjectToSolution {
     
-    public static init(uri: vscode.Uri, extensionUri: vscode.Uri) {
+    public static init(uri: vscode.Uri, context: vscode.ExtensionContext) {
             vscode.window.showInputBox({ignoreFocusOut: true, prompt: 'Type the project name', value: 'New project name'})
                 .then((newFileName) => {
                     if (typeof(newFileName) === undefined || newFileName === '') {
                         vscode.window.showErrorMessage('Please input a valid name or press Scape to cancel the operation!');
-                        return this.init(uri, extensionUri); 
+                        return this.init(uri, context); 
                     }
                     
                     // Acquiring the solution root folder
@@ -29,7 +28,7 @@ export class AddProjectToSolution {
                     // Verify if project already exist
                     if (fs.existsSync(newFilePath)) {
                         vscode.window.showErrorMessage(`Project ${newFileName} already exist`);
-                        return this.init(uri, extensionUri);    
+                        return this.init(uri, context);    
                     }
                     
                     let rootDir = getProjectRootDirOrFilePath(root);
@@ -41,12 +40,14 @@ export class AddProjectToSolution {
                     
                     // Create a project
                     // TODO: Pass project solution, project name
-                    const projectState: ProjectState = ProjectState.Add;
-                    CreateProjectPanel.createOrShow(extensionUri, projectState);
+                    if (newFileName) {
+                        const panel = new Panel(context, newFileName, rootDir, "Add Project", {folder: 'media', file: 'addProjectIcon.png'});
+                        panel.webViewPanel?.onDidDispose(()=> {panel.webViewPanel = undefined;}, context.subscriptions);
+                        
+                    };
                 }
             ); 
     }
-    
 }
 
 // function to detect the root directory where the .csproj is included
@@ -54,7 +55,6 @@ function getProjectRootDirOrFilePath(filePath: any){
     const paths : string[] = filePath.split("\\");
     const solution: string = filePath + path.sep + paths[paths.length-1] + '.sln';
     
-
     if (!fs.existsSync(solution)) {
         return null;
     }