+++ CHANGELOG.md
@@ -5,6 +5,17 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Released]
 
+## [1.2.4] - 2021-08-01
+
+- **_Fix_**: Solution was being created with project name rather than solution data from solution field.
+- **_New Features added_**:
+- **_Add Project to a Solution_** : Capability to add projects to the same solution with a click of a button. You can select a different project framework as well as the template.
+- **_Submenu With Options_** :
+- Create Class
+- Create Interface
+- Create Record
+- Create Struct
+
 ## [1.2.3] - 2021-07-18
 
 - **_Fix_**: .NET target frameworks list on project creation are based on OS and SDKs installed.